Move address to root directory
