+++ collector/builder/builder.go
@@ -2,9 +2,9 @@ package builder
 
 import (
 	"errors"
+	address2 "github.com/nervosnetwork/ckb-sdk-go/address"
 	"github.com/nervosnetwork/ckb-sdk-go/collector"
 	"github.com/nervosnetwork/ckb-sdk-go/collector/handler"
-	"github.com/nervosnetwork/ckb-sdk-go/script/address"
 	"github.com/nervosnetwork/ckb-sdk-go/transaction"
 	"github.com/nervosnetwork/ckb-sdk-go/types"
 	"reflect"
@@ -86,7 +86,7 @@ func (r *SimpleTransactionBuilder) AddOutput(output *types.CellOutput, data []by
 }
 
 func (r *SimpleTransactionBuilder) AddOutputByAddress(addr string, capacity uint64) error {
-	a, err := address.Decode(addr)
+	a, err := address2.Decode(addr)
 	if err != nil {
 		return err
 	}
